+++ app/view_models/__init__.py
@@ -1,0 +1,1 @@
+from app.view_models.items import ItemsViewModel
