--- view_models/__init__.py
@@ -1,1 +1,0 @@
-from view_models.items import ItemsViewModel
